zksync_vm2

Module addressing_modes

Addressing modes supported by EraVM.

Structs

AbsoluteStack

Absolute addressing into stack.

AdvanceStackPointer

Same as RelativeStack, but moves the stack pointer on access (decreases it when reading data; increases when writing data).

Arguments

Arguments provided to an instruction in an EraVM bytecode.

CodePage

Absolute addressing into the code page of the currently executing program.

Immediate1

Immediate value passed as a first instruction arg.

Immediate2

Immediate value passed as a second instruction arg.

NotRegisterOrImmediate

Error converting AnySource to RegisterOrImmediate.

Register

Representation of one of 16 VM registers.

Register1

Register passed as a first instruction argument.

Register2

Register passed as a second instruction argument.

RegisterAndImmediate

Combination of a register and an immediate value wrapped by AbsoluteStack, RelativeStack, AdvanceStackPointer and CodePage addressing modes.

RelativeStack

Relative addressing into stack (relative to the VM stack pointer).

Enums

AnyDestination

All supported addressing modes for the first destination argument.

AnySource

All supported addressing modes for the first source argument.

RegisterOrImmediate

Register or immediate addressing modes required by some VM instructions.