Crate zksync_vm2

High-Performance ZKsync Era VM

This crate provides high-performance VirtualMachine for ZKsync Era.

Re-exports

pub use zksync_vm2_interface as interface;

Modules

addressing_modes

Addressing modes supported by EraVM.

precompiles

Precompiles support.

testonly

Test-only tools for EraVM.

Structs

FatPointer

Fat pointer to a heap location.

Instruction

Single EraVM instruction (an opcode + Arguments).

ModeRequirements

VM execution mode requirements (kernel only, not in static call) that can be placed on instructions.

Program

Compiled EraVM bytecode.

Settings

VirtualMachine settings.

Snapshot

Opaque snapshot of a WorldDiff output by its eponymous method. Can be provided to WorldDiff::events_after() etc. to get data after the snapshot was created.

StorageChange

Change in a single storage slot.

StorageSlot

Storage slot information returned from StorageInterface::read_storage().

VirtualMachine

High-performance out-of-circuit EraVM implementation.

WorldDiff

Pending modifications to the global state that are executed at the end of a block. In other words, side effects.

Enums

ExecutionEnd

VM stop reason returned from VirtualMachine::run().

Predicate

Predicate for an instruction. Encoded so that comparing it to flags is efficient.

Traits

StorageInterface

VM storage access operations.

World

Encapsulates VM interaction with the external world. This includes VM storage and decomitting (loading) bytecodes for execution.