Struct zksync_vm2::VirtualMachine

pub struct VirtualMachine<T, W> { /* private fields */ }

High-performance out-of-circuit EraVM implementation.

Implementations

impl<T: Tracer, W: World<T>> VirtualMachine<T, W>

pub fn new( address: H160, program: Program<T, W>, caller: H160, calldata: &[u8], gas: u32, settings: Settings, ) -> Self

Creates a new VM instance.

pub fn world_diff(&self) -> &WorldDiff

Provides a reference to the World diff accumulated by VM execution so far.

pub fn run(&mut self, world: &mut W, tracer: &mut T) -> ExecutionEnd

Runs this VM with the specified World and Tracer until an end of execution due to a hook, or an error.

pub fn resume_with_additional_gas_limit( &mut self, world: &mut W, tracer: &mut T, gas_limit: u32, ) -> Option<(u32, ExecutionEnd)>

Returns how much of the extra gas limit is left and the stop reason, unless the extra gas limit was exceeded.

Needed to support account validation gas limit. We cannot simply reduce the available gas, as contracts might behave differently depending on remaining gas.

pub fn make_snapshot(&mut self)

Creates a VM snapshot. The snapshot can then be rolled back to, or discarded.

Panics

Panics if called outside the initial (bootloader) callframe.

Panics if this VM already has a snapshot.

pub fn rollback(&mut self)

Returns the VM to the state it was in when Self::make_snapshot() was called.

Panics

Panics if this VM doesn’t hold a snapshot.

Panics if called outside the initial (bootloader) callframe.

pub fn pop_snapshot(&mut self)

Pops a previously made snapshot without rolling back to it. This effectively commits all changes made up to this point, so that they cannot be rolled back.

Panics

Panics if called outside the initial (bootloader) callframe.