Struct zksync_vm2::WorldDiff

pub struct WorldDiff { /* private fields */ }

Pending modifications to the global state that are executed at the end of a block. In other words, side effects.

Implementations

impl WorldDiff

pub fn storage_refunds(&self) -> &[u32]

Returns recorded refunds for all storage operations.

pub fn pubdata_costs(&self) -> &[i32]

Returns recorded pubdata costs for all storage operations.

pub fn get_storage_changes( &self, ) -> impl Iterator<Item = ((H160, U256), StorageChange)> + '_

Gets changes for all touched storage slots.

pub fn get_storage_changes_after( &self, snapshot: &Snapshot, ) -> impl Iterator<Item = ((H160, U256), StorageChange)> + '_

Gets changes for storage slots touched after the specified snapshot was created.

pub fn events_after(&self, snapshot: &Snapshot) -> &[Event]

Returns events emitted after the specified snapshot was created.

pub fn l2_to_l1_logs_after(&self, snapshot: &Snapshot) -> &[L2ToL1Log]

Returns L2-to-L1 logs emitted after the specified snapshot was created.

pub fn decommitted_hashes(&self) -> impl Iterator<Item = U256> + '_

Returns hashes of decommitted contract bytecodes in no particular order. Note that this includes failed (out-of-gas) decommitments.

pub fn snapshot(&self) -> Snapshot

Get a snapshot for selecting which logs & co. to output using Self::events_after() and other methods.